:recycle: separate file about ViewGraph+SetRect
